Separate package target directory from nuspec base path
When building multiple projects it is reasonable to assume that they will
each have their own TargetDir, from which files should be packaged.
However it is useful to be able to specify a single output directory for
all built packages by overriding the OutDir.

This change maintains compatibility with existing build scripts which use
NugetPackageTargetDir to mean both the source of files and the package's
destination.
New scripts may override the NuSpecBasePath to specify the source location
for packaged files, and override the package destination by specifying
NugetPackageOutDir.

NugetPackageTargetDir should probably be considered deprecated.

Refers to #17